Add comparative benchmark vs solana-program-test (#4)
* add counter test

* lint

* add comparative benchmark

* remove old code

+++ lite-program-test/tests/counter/src/state.rs
@@ -2,5 +2,5 @@ use borsh::{BorshDeserialize, BorshSerialize};
 
 #[derive(BorshSerialize, BorshDeserialize, Debug, Clone)]
 pub struct Counter {
-    pub count: u64,
+    pub count: u32,
 }
